refactor(auth): merge store.type into auth.type and simplify store
