+++ packages/shared/src/features/auth/types/index.ts
@@ -1,2 +1,1 @@
 export * from './auth.type';
-export * from './store.type';
